Add connection property sqlVariantAsNull
Add a boolean connection property sqlVariantAsNull that instructs the
driver to interpret SQL_VARIANT columns as null values. Previously the
driver had no handling for SQL_VARIANT columns and would close any
ResultSet that included a column of type SQL_VARIANT. With this option
enabled, the driver will skip over those column values but otherwise
allow ResultSets to include SQL_VARIANT columns.

This option is disabled by default to allow the existing behavior of
throwing an exception if a SQL_VARIANT is encountered to continue.